refactor: Move `createQueryClient` to tRPC directory
